iOS-33: fixed warning: use same type in overriden implementation

+++ Console/Classes/RoundRobinException.m
@@ -23,7 +23,7 @@
 
 @implementation RoundRobinException
 
-+ (NSString *)controller:(ORControllerConfig *)controller exceptionMessageOfCode:(int)code {
++ (NSString *)controller:(ORControllerConfig *)controller exceptionMessageOfCode:(NSInteger)code {
 	NSString *errorMessage = nil;
 	if (code != 200) {
 		switch (code) {
@@ -38,7 +38,7 @@
 				break;
 		}
 		if (!errorMessage) {
-			errorMessage = [NSString stringWithFormat:@"Occured unknown error, satus code is %d", code];
+            errorMessage = [NSString stringWithFormat:@"Occured unknown error, satus code is %ld", (long)code];
 		}
 	}
 	return errorMessage;
